Added Lab 2
In this lab we practiced data cleaning, data transformations,  we practiced data augmentation. All using a data set for avocado sales we were interested in things like sales by region, max and average sales in particular regions. We practiced pivoting with untidy data to make it more tidy. We also did more plotting and tried to uncover interesting facts about the data in the process.