Penalized Regression Practice
In this activity we practice methods of regression that will bias the coefficients in order to make better predictions on new data. We practice Ridge, Lasso and their combination which is Elastic Net regression to showcase what changes in the coefficients and how the Metrics with which we evaluate the models change as a result of these modified regression models.