MultiClass Models
Here we practice using KNN Decision Trees and Linear Discriminant Analysis in order to do classification with multiclass problems. Specifically we emphasised the differences between one vs rest and one vs one , the nuances between what type of metrics will be best at judging the models and what the intuition is for what models to fit.